% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_wombling.R
\name{bayes_cwomb}
\alias{bayes_cwomb}
\title{Curvilinear Bayesian Wombling}
\usage{
bayes_cwomb(
  coords = NULL,
  model = NULL,
  cov.type = c("gaussian", "matern1", "matern2"),
  nbatch = NULL,
  curve = NULL,
  type = c("rectilinear", "riemann.sum"),
  approx = NULL,
  verbose = T,
  rule.len = 10
)
}
\arguments{
\item{coords}{coordinates for observed process (order \eqn{L} x \eqn{2})}

\item{model}{the posterior samples from the MCMC fit}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{nbatch}{number of batches}

\item{curve}{coordinates for the curve}

\item{type}{specify type or rule of quadrature}

\item{approx}{if TRUE curve will be approximated}

\item{verbose}{if TRUE will print status}

\item{rule.len}{specifies accuracy for quadrature}
}
\description{
performs Bayesian wombling on curves within the spatial surface.
}
