% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezierCurve.R
\name{bezierCurve}
\alias{bezierCurve}
\title{Bezier Splines}
\usage{
bezierCurve(x, y, n = 10)
}
\arguments{
\item{x}{x-coordinate for points}

\item{y}{y-coordinate for points}

\item{n}{number of points to interpolate}
}
\description{
A function for computing Bezier splines from a set of annotated points
}
\keyword{spWombling}
