% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_gaussian.R
\name{cov_gaussian}
\alias{cov_gaussian}
\title{Gaussian covariance kernel}
\usage{
cov_gaussian(Delta = NULL, phi = NULL, sig2 = NULL)
}
\arguments{
\item{Delta}{distance matrix}

\item{phi}{spatial range}

\item{sig2}{spatial variance}
}
\description{
Gaussian covariance kernel
}
\keyword{cov_gaussian}
