% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern.R
\name{cov_matern}
\alias{cov_matern}
\title{Mat\'ern(\eqn{\nu}) covariance kernel}
\usage{
cov_matern(Delta = NULL, phi = NULL, sig2 = NULL, nu = NULL)
}
\arguments{
\item{Delta}{distance matrix}

\item{phi}{spatial range}

\item{sig2}{spatial variance}

\item{nu}{fractal parameter}
}
\description{
Mat\'ern(\eqn{\nu}) covariance kernel
}
\keyword{cov_matern1}
