% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_matern1.R
\name{cov_matern1}
\alias{cov_matern1}
\title{Mat\'ern(\eqn{\nu=3/2}) covariance kernel}
\usage{
cov_matern1(Delta = NULL, phi = NULL, sig2 = NULL)
}
\arguments{
\item{Delta}{distance matrix}

\item{phi}{spatial range}

\item{sig2}{spatial variance}
}
\description{
Mat\'ern(\eqn{\nu=3/2}) covariance kernel
}
\keyword{cov_matern1}
