% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_summary.R
\name{hlm_summary}
\alias{hlm_summary}
\title{Model Summary for spatial hierarchical Bayes models}
\usage{
hlm_summary(
  chain = NULL,
  nburn = NULL,
  niter = NULL,
  thin = NULL,
  latent_inf = T
)
}
\arguments{
\item{chain}{MCMC posterior samples}

\item{nburn}{number of burn-in samples}

\item{niter}{number of MCMC iterations}

\item{thin}{thinning interval}

\item{latent_inf}{if true returns inference for spatial effects \eqn{Z}}
}
\description{
Provides a model summary for fitted hlmBayes_sp models
}
