% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmBayes_sp.R
\name{hlmBayes_sp}
\alias{hlmBayes_sp}
\title{A spatial hierarchical Bayes Markov Chain Monte Carlo sampler}
\usage{
hlmBayes_sp(
  y = NULL,
  coords = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  a.sigma = NULL,
  b.sigma = NULL,
  a.tau = NULL,
  b.tau = NULL,
  lower.phi = NULL,
  upper.phi = NULL,
  cov.type = NULL,
  verbose = TRUE,
  trgtFn.compute = FALSE,
  digits = 3
)
}
\arguments{
\item{y}{observed response (order \eqn{L} x  \eqn{1})}

\item{coords}{coordinates for observed process (order \eqn{L} x \eqn{2})}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in samples}

\item{report}{batch length}

\item{a.sigma}{shape parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{b.sigma}{scale parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{a.tau}{shape parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{b.tau}{scale parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{lower.phi}{lower limit for uniform prior on \eqn{\phi}}

\item{upper.phi}{upper limit for uniform prior on \eqn{\phi}}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{verbose}{if true prints output for batches}

\item{trgtFn.compute}{compute posterior}

\item{digits}{rounding digits}
}
\description{
Fits a univariate Gaussian spatial regression model: \eqn{Y(s)=x(s)^T\beta+Z(s)+\epsilon}. Parameters not listed are optional.
}
\keyword{spWombling}
