% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meuse.R
\docType{data}
\name{meuse}
\alias{meuse}
\title{Meuse river data set}
\format{
Refer to sp::meuse for full documentation
}
\source{
sp
}
\usage{
data(meuse)
}
\description{
This data set gives locations and topsoil heavy metal concentrations, along with a number of soil and landscape variables at the observation locations, collected in a flood plain of the river Meuse, near the village of Stein (NL). Heavy metal concentrations are from composite samples of an area of approximately 15 m x 15 m.
}
\keyword{datasets}
