% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETemp.R
\docType{data}
\name{NETemp.dat}
\alias{NETemp.dat}
\title{Monthly weather station temperature data across the Northeastern US}
\format{
A data frame containing 356 rows (weather stations) and 132 columns.
}
\source{
spBayes
}
\usage{
data(NETemp.dat)
}
\description{
Monthly temperature data (Celsius) recorded across the Northeastern US starting in January 2000.
Station UTM coordinates and elevation are also included.
}
\keyword{datasets}
