% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plot.R
\name{sp_plot}
\alias{sp_plot}
\title{Spatial Plots}
\usage{
sp_plot(
  col.seq.length = NULL,
  col.text = NULL,
  data_frame = NULL,
  shape = NULL,
  zlim = NULL,
  xlab = NULL,
  ylab = NULL,
  lab = NULL,
  add = FALSE,
  categorical = FALSE,
  points.plot = FALSE,
  points.cex = 0.3,
  contour.plot = FALSE,
  legend = TRUE,
  digits = "\%2f",
  raster.surf = FALSE,
  extend = FALSE,
  sig = NULL,
  grid = FALSE
)
}
\arguments{
\item{col.seq.length}{length of colors in color pallate}

\item{col.text}{names of color pallate}

\item{data_frame}{a matrix of coordinates and response (order \eqn{L} x  \eqn{3})}

\item{shape}{shape file for respone if available}

\item{zlim}{upper limit for response (optional)}

\item{xlab}{(Plotting parameter) label for x-axis}

\item{ylab}{(Plotting parameter) label for y-axis}

\item{lab}{(Plotting parameter) label for z-axis}

\item{add}{if plot needs to be added to an existing plot}

\item{categorical}{if response is cateforgical}

\item{points.plot}{if true plots points}

\item{contour.plot}{if true plots contours}

\item{legend}{if true plots legend}

\item{raster.surf}{if true returns a raster object of surface}

\item{sig}{vector of colors indicating significance for points (order \eqn{L} x \eqn{1})}

\item{grid}{if true plots grid}
}
\description{
A function that plots a kriged spatial field from coordinates.
}
