% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsample.R, R/spsample_beta.R
\name{spsample}
\alias{spsample}
\title{Posterior samples for spatial effects and regression coefficients}
\usage{
spsample(
  y = NULL,
  X = NULL,
  coords = NULL,
  phi = NULL,
  sig2 = NULL,
  tau2 = NULL,
  cov.type = c("matern1", "matern2", "gaussian"),
  silent = TRUE
)

spsample(
  y = NULL,
  X = NULL,
  coords = NULL,
  phi = NULL,
  sig2 = NULL,
  tau2 = NULL,
  cov.type = c("matern1", "matern2", "gaussian"),
  silent = TRUE
)
}
\arguments{
\item{y}{response}

\item{X}{design matrix}

\item{coords}{spatial co-ordinates}

\item{phi}{posterior samples of the spatial range parameter}

\item{sig2}{posterior samples of the spatio-temporal variance parameter (partial sill)}

\item{tau2}{posterior samples of the error variance (nugget)}

\item{cov.type}{type of covariance kernel being used. Choices include Exponential, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2}), Gaussian}

\item{silent}{logical argument for print-statements}

\item{t}{temporal coordinates}

\item{phis}{posterior samples of the spatial range parameter}

\item{phit}{posterior samples of the temporal range parameter}
}
\description{
Posterior sampling algorithm for generating samples of spatial random effects and
regression coefficients using a Gibbs sampling algorithm.

Posterior sampling algorithm for generating samples of spatio-temporal random effects and
regression coefficients using a Gibbs sampling algorithm.
}
\keyword{spsample}
\keyword{spsample_beta}
