% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_summary.R
\name{hlm_summary}
\alias{hlm_summary}
\title{Model Summary for spatial hierarchical Bayes models}
\usage{
hlm_summary(sig2 = NULL, tau2 = NULL, phi = NULL, beta = NULL, z = NULL)
}
\arguments{
\item{sig2}{MCMC posterior samples for \eqn{\sigma^2}}

\item{tau2}{MCMC posterior samples for \eqn{\tau^2}}

\item{phi}{MCMC posterior samples for \eqn{\phi}}

\item{beta}{MCMC posterior samples for \eqn{\beta}}

\item{z}{MCMC posterior samples for \eqn{Z}}
}
\description{
Provides a model summary for fitted hlmBayes_sp models
}
\keyword{hlm_summary}
