% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmBayes_sp.R
\name{hlmBayes_sp}
\alias{hlmBayes_sp}
\title{A spatial hierarchical Bayes Markov Chain Monet Carlo sampler}
\usage{
hlmBayes_sp(
  coords = NULL,
  y = NULL,
  X = NULL,
  z_init = NULL,
  D = NULL,
  phis_init = NULL,
  lower_phis = NULL,
  upper_phis = NULL,
  sigma2_init = NULL,
  shape_sigma = NULL,
  scale_sigma = NULL,
  tau2_init = NULL,
  shape_tau = NULL,
  scale_tau = NULL,
  beta_init = NULL,
  mean_beta = NULL,
  prec_beta = NULL,
  steps_init = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  cov.type = c("exponential", "gaussian", "matern1", "matern2"),
  spatial.rand.effect = T,
  replication = F,
  verbose = T,
  nu_est = F,
  stepnu_init = NULL,
  nu_init = NULL,
  lower_nu = NULL,
  upper_nu = NULL
)
}
\arguments{
\item{coords}{coordinates for observed process (order \eqn{L} x \eqn{2})}

\item{y}{observed response (order \eqn{L} x  \eqn{1})}

\item{X}{a matrix of covariates (order \eqn{L} x  \eqn{p})}

\item{z_init}{starting values of spatial effects (order \eqn{L} x  \eqn{1})}

\item{D}{(use if replication at co-ordinate level) index for observations}

\item{phis_init}{starting value for \eqn{\phi}}

\item{lower_phis}{lower bound for uniform prior on \eqn{\phi}}

\item{upper_phis}{upper bound for uniform prior on \eqn{\phi}}

\item{sigma2_init}{starting value for \eqn{\phi}}

\item{shape_sigma}{lower bound for uniform prior on \eqn{\phi}}

\item{scale_sigma}{upper bound for uniform prior on \eqn{\phi}}

\item{tau2_init}{starting value for \eqn{\phi}}

\item{shape_tau}{lower bound for uniform prior on \eqn{\phi}}

\item{scale_tau}{upper bound for uniform prior on \eqn{\phi}}

\item{beta_init}{starting value for \eqn{\phi}}

\item{mean_beta}{lower bound for uniform prior on \eqn{\phi}}

\item{prec_beta}{upper bound for uniform prior on \eqn{\phi}}

\item{steps_init}{tuning parameter for \eqn{\phi}}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in samples}

\item{report}{batch length}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{spatial.rand.effect}{if false perform linear regression}

\item{replication}{refer to parameter D above}

\item{verbose}{if true prints output for batches}

\item{nu_est}{(if cov.type not specified turns to true) smoothness parameter for Mat\'ern}

\item{stepnu_init}{tuning parameter for \eqn{\nu}}

\item{nu_init}{starting value for \eqn{\nu}}

\item{lower_nu}{lower bound for uniform prior on \eqn{\nu}}

\item{upper_nu}{upper bound for uniform prior on \eqn{\nu}}
}
\description{
Fits a univariate Gaussian spatial regression model: \eqn{Y(s)=x(s)^T\beta+Z(s)+\epsilon}. Parameters not listed are optional.
}
