% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_gradient.R
\name{spatial_gradient}
\alias{spatial_gradient}
\title{Gradient and Curvature Assessment}
\usage{
spatial_gradient(
  coords = NULL,
  chain = NULL,
  cov.type = c("gaussian", "matern1", "matern2"),
  grid.points = NULL,
  samples = NULL,
  niter = NULL,
  nbatch = 200,
  nburn = NULL,
  return.mcmc = TRUE
)
}
\arguments{
\item{coords}{coordinates for observed process (order \eqn{L} x \eqn{2})}

\item{chain}{the posterior samples from the MCMC fit}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{grid.points}{coordinates for grid over observed process (order \eqn{n_G} x \eqn{2})}

\item{samples}{index of samples from the chain to be used}

\item{niter}{number of MCMC iterations}

\item{nbatch}{number of batches}

\item{nburn}{number of burn-in samples}

\item{return.mcmc}{if true returns mcmc samples of gardients}
}
\description{
performs gradient and curvature assessment on the estimated surface from a hierarchcial Bayesian spatial model. Parameters not listed are optional
}
